rename: 오타가 있는 dropdown 폴더명 변경
